Move names to a more appropriate package.
